Implemented waiting until file is deleted or time out.

+++ messengerbot/servicehandlers/Service.py
@@ -110,6 +110,8 @@ class Service(object):
         image_file_path = os.path.join(LocalConfigChecker.program_directory, "temp_text_image.png")
         image_text = message.message_body
 
+        self.wait_until_delete(image_file_path, 5)
+
         padding = (2, 2)
         fontsize = 30
         font_file = get_font("NotCourierSans.otf")
@@ -243,8 +245,31 @@ class Service(object):
             :return: None
             """
             time.sleep(after_seconds)
-            os.remove(file_path)
+            try:
+                os.remove(file_path)
+            except FileNotFoundError:
+                pass
 
         delete_thread = Thread(target=delete_file_and_wait)
         delete_thread.daemon = True
         delete_thread.start()
+
+    @staticmethod
+    def wait_until_delete(file_path: str, time_out: int) -> None:
+        """
+        Waits until a file is deleted. If it is not deleted within a given time period, manually delete the file
+
+        :param file_path: the file to wait for
+        :param time_out: the maximum time period to wait
+        :return: None
+        """
+        sleep_seconds = 0
+        while os.path.isfile(file_path):
+            time.sleep(1)
+            sleep_seconds += 1
+
+            if sleep_seconds >= time_out:
+                try:
+                    os.remove(file_path)
+                except FileNotFoundError:
+                    pass
